1. Moved Environment class back to the systemtest project 2. Moven methods to configure docker to StUtils
