+++ systemtest/src/main/java/io/strimzi/systemtest/MessagingBaseST.java
@@ -9,7 +9,6 @@ import io.strimzi.systemtest.clients.api.ClientArgument;
 import io.strimzi.systemtest.clients.api.ClientArgumentMap;
 import io.strimzi.systemtest.clients.api.MsgCliApiClient;
 import io.strimzi.systemtest.clients.api.VerifiableClient;
-import io.strimzi.test.Environment;
 import io.strimzi.test.TestUtils;
 import io.vertx.core.json.JsonObject;
 import org.apache.logging.log4j.LogManager;
