+++ systemtest/src/main/java/io/strimzi/systemtest/Resources.java
@@ -59,7 +59,6 @@ import io.strimzi.api.kafka.model.KafkaUserScramSha512ClientAuthentication;
 import io.strimzi.api.kafka.model.KafkaUserTlsClientAuthentication;
 import io.strimzi.api.kafka.model.SingleVolumeStorage;
 import io.strimzi.systemtest.utils.StUtils;
-import io.strimzi.test.Environment;
 import io.strimzi.test.TestUtils;
 import io.strimzi.test.k8s.KubeClient;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
@@ -77,7 +76,7 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.IntStream;
 
-import static io.strimzi.test.TestUtils.changeOrgAndTag;
+import static io.strimzi.systemtest.utils.StUtils.changeOrgAndTag;
 import static io.strimzi.test.TestUtils.toYamlString;
 
 @SuppressWarnings({"checkstyle:ClassDataAbstractionCoupling", "checkstyle:ClassFanOutComplexity"})
@@ -244,8 +243,8 @@ public class Resources extends AbstractResources {
     }
 
     public KafkaBuilder defaultKafka(String name, int kafkaReplicas, int zookeeperReplicas) {
-        String tOImage = TestUtils.changeOrgAndTag(getImageValueFromCO("STRIMZI_DEFAULT_TOPIC_OPERATOR_IMAGE"));
-        String uOImage = TestUtils.changeOrgAndTag(getImageValueFromCO("STRIMZI_DEFAULT_USER_OPERATOR_IMAGE"));
+        String tOImage = StUtils.changeOrgAndTag(getImageValueFromCO("STRIMZI_DEFAULT_TOPIC_OPERATOR_IMAGE"));
+        String uOImage = StUtils.changeOrgAndTag(getImageValueFromCO("STRIMZI_DEFAULT_USER_OPERATOR_IMAGE"));
 
         return new KafkaBuilder()
                     .withMetadata(new ObjectMetaBuilder().withName(name).withNamespace(client().getNamespace()).build())
@@ -658,9 +657,9 @@ public class Resources extends AbstractResources {
                     break;
                 default:
                     if (envVar.getName().contains("STRIMZI_DEFAULT")) {
-                        envVar.setValue(TestUtils.changeOrgAndTag(envVar.getValue()));
+                        envVar.setValue(StUtils.changeOrgAndTag(envVar.getValue()));
                     } else if (envVar.getName().contains("IMAGES")) {
-                        envVar.setValue(TestUtils.changeOrgAndTagInImageMap(envVar.getValue()));
+                        envVar.setValue(StUtils.changeOrgAndTagInImageMap(envVar.getValue()));
                     }
             }
         }
@@ -676,7 +675,7 @@ public class Resources extends AbstractResources {
                     .editTemplate()
                         .editSpec()
                             .editFirstContainer()
-                                .withImage(TestUtils.changeOrgAndTag(coImage))
+                                .withImage(StUtils.changeOrgAndTag(coImage))
                             .endContainer()
                         .endSpec()
                     .endTemplate()
